% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepFilterIndependence.R
\name{parfilter}
\alias{parfilter}
\title{Par-filter}
\usage{
parfilter(
  p,
  error_targets,
  u,
  groups = NULL,
  u_groups = NULL,
  w = NULL,
  selections = NULL,
  adaptive,
  lambda = NULL,
  auto = FALSE,
  omega = 0.5,
  group_options = NULL,
  partition = NULL,
  method = "Simes"
)
}
\arguments{
\item{p}{a matrix of p-values}

\item{error_targets}{numeric of FDR targets. The first \eqn{n} elements are the study-specific FDR. The last element is the replicability FDR.}

\item{u}{the replicability threshold}

\item{groups}{the partition of the n studies. It is a list of numerics denoting which studies belong to which groups. This can be left NULL if \code{auto = TRUE} or a type of \code{partition} is specified.}

\item{u_groups}{numeric of the replicability thresholds for each group. If left NULL, \code{u_groups} will automatically be generated.}

\item{w}{numeric of error weights. If left NULL, \code{w} will automatically be generated.}

\item{selections}{list of numerics denoting the selections. If left NUll, \code{selections} will be automatically be generated based on whether \code{adaptive} is \code{TRUE} or \code{FALSE}.}

\item{adaptive}{Boolean indicating whether to use adaptivity or not.}

\item{lambda}{numeric of tuning parameters for adaptivity. If left NUll, \code{lambda} will be automatically be generated based on whether \code{error_targets}.}

\item{auto}{Boolean indicating whether to automatically compute the study partition}

\item{omega}{the overfitting tuning parameter for the automatic partioning algorithm}

\item{group_options}{list of candidate partitions(a list) for the automatic partitioning algorithm}

\item{partition}{the type of partition if groups is NULL. If \code{partition = "minimum"}, a neutral partition will be created with \eqn{K = 2} groups. If \code{partition = "minimum"}, a neutral partition will be created with \eqn{K = u} groups.}

\item{method}{the combining method}
}
\value{
A list containing the study-specific rejections, the rejections for assessing replicability,
 a numeric of study-specific and group-specific FDR target levels, the partition, the local replicability levels, and the local error weights.
}
\description{
Par-filter
}
\examples{
obj <- parfilter(p = DMD.pvalues, error_targets = rep(0.05, 5),
u = 3, groups = list(c(1,3),c(2,4)),
group_options = c(2,1),
selections = NULL, u_groups = c(2,1),
adaptive = TRUE, lambda = NULL,
auto = FALSE, omega = 0.75, w = c(0.5,0.5),
partition = NULL, method = "Simes")
print(obj)
}
