% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParFilter functions.R
\name{ParFilter_FDR}
\alias{ParFilter_FDR}
\title{Title}
\usage{
ParFilter_FDR(
  p_mat,
  X_list,
  u,
  q,
  K,
  method,
  adaptive = TRUE,
  lambdas = rep(0.5, K),
  cross_weights = FALSE,
  inflate = FALSE
)
}
\arguments{
\item{p_mat}{mxn matrix of p-values}

\item{X_list}{List of covariates (numeric of matrix) for each study}

\item{u}{Replicability level}

\item{q}{FDR target}

\item{K}{Number of groups}

\item{method}{Method for combining p-values: "Fisher", "Stouffer", or "Simes"}

\item{adaptive}{logical indicating whether to use adaptive weighted null proportion estimates or not}

\item{lambdas}{numeric of tuning parameters for the weighted null proportion estimates.}

\item{cross_weights}{Set as TRUE if the p-values are dependent within studies, otherwise leave it as FALSE.}

\item{inflate}{Use a harmonic number as the weighted null proportion estimates.}
}
\value{
The rejections for determining u/[n] replicability
}
\description{
Title
}
