% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stirlingfunctions.R
\name{StirlingPartitions}
\alias{StirlingPartitions}
\title{Generate Partitions}
\usage{
StirlingPartitions(n, k)
}
\arguments{
\item{n}{number of studies}

\item{k}{maximum number of non-empty sets in the partition}
}
\value{
A list of partitions of \eqn{1,...,n} of up to k non-empty subsets.
}
\description{
Generate Partitions
}
\examples{
StirlingPartitions(4,3)
}
