% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParFilter functions.R
\name{ParFilter_FDR}
\alias{ParFilter_FDR}
\title{ParFilter}
\usage{
ParFilter_FDR(
  p_mat,
  X_list,
  u,
  q,
  K,
  method,
  adaptive = TRUE,
  lambdas = rep(0.5, K),
  cross_weights = FALSE
)
}
\arguments{
\item{p_mat}{mxn matrix of p-values.}

\item{X_list}{list of length n, containing the covariates for each study.}

\item{u}{replicability threshold.}

\item{q}{FDR target.}

\item{K}{number of groups. ParFilter_FDR will automatically partition the n studies in two K groups of approximately equal sizes.}

\item{method}{combining function for creating the local GBHPC p-values. Can be either: "Fisher", "Stouffer", or "Simes".}

\item{adaptive}{logical indicating whether to use adaptive null proportion estimates or not.}

\item{lambdas}{set as TRUE if the p-values are dependent within studies, otherwise leave it as FALSE.}

\item{cross_weights}{numeric of tuning parameters for the null proportion estimates.}
}
\value{
returns a numeric of features indices to be rejected.
}
\description{
ParFilter
}
